Refactor my code and apply reviewer suggestions
