+++ src/components/Books/BookList.js
@@ -1,7 +1,7 @@
 import React, { useEffect } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
 
-import { fetchBooks } from '../../redux/books/bookActions';
+import { fetchBooks } from '../../redux/books/books';
 
 import Book from './Book';
