+++ src/index.js
@@ -2,12 +2,9 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import { BrowserRouter } from 'react-router-dom';
 import { Provider } from 'react-redux';
-import axios from 'axios';
 
 import store from './redux/configureStore';
 import App from './App';
-
-axios.defaults.baseURL = 'https://us-central1-bookstore-api-e63c8.cloudfunctions.net/bookstoreApi/apps/eZNPE2dPLtWHMUhnsTBH/books';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
